fix compiler warning

+++ iMEGA/Camera uploads/UploadManagers/CameraUploadManager.m
@@ -153,7 +153,7 @@ static const NSInteger MaxConcurrentVideoOperationCount = 1;
 - (void)uploadNextAssetsWithNumber:(NSInteger)number mediaType:(PHAssetMediaType)mediaType {
     NSArray *records = [CameraUploadRecordManager.shared fetchNonUploadedRecordsWithLimit:number mediaType:mediaType error:nil];
     if (records.count == 0) {
-        MEGALogDebug(@"[Camera Upload] no more local asset to upload for media type %li", mediaType);
+        MEGALogDebug(@"[Camera Upload] no more local asset to upload for media type %li", (long)mediaType);
         return;
     }
